Handle casacore dependency correctly
The FindCasacore.cmake module under the cmake/ directory seems pretty
old (2011), and was clearly coded against an old version of casacore.
This is made evident by the way headers are looked for: they don't
contain the "casacore/" prefix, and instead they are looked for both
under include/ and include/casacore/. This is from a time when the
transition to fully-prefixed headers was happening, and users could
still choose from both. The C++ code in this project correctly includes
casacore headers with the full prefix though, and hence the
FindCasacore.cmake module should look for "casacore/"-prefixed headers
only inside include/.

This commit changes how casacore headers are looked for by the cmake
machinery, following the guidelines above.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>