Unify CUDA option declaration and enabling
The previous way in which CUDA was looked for, enabled, and presented to
the user as an option was a bit confusing, and didn't have a clear
if-then narrative.

This commit asssumes CUDA support is optional at compile time. An option
is presented to the user to enable/disable this support (it's ON by
default), and whenever support is requested and not found an error is
raised telling users how to disable it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>