Make PhaseRotateTests compilable in non-CUDA builds
The previous incarnation of the PhaseRotateTests assumed that CUDA was
always available at compile time, but this will not always be the case.
When not present we want to avoid compiling any CUDA-specific code and
tests.

This commit changes the useCuda runtime parameter used by the tests in
the PhaseRotateTests to a compile-time argument (i.e., a template value)
that is used to choose between the CPU and CUDA version of the test.
This required some modification to the code, but nothing too bad. In
turn, it allowed us to include certain test code in the compilation when
USE_CUDA is defined.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>