Improve warning flag handling
There were two problems with how warning flags were added: first, the
quoting was wrong and things broke if users specified any value for
CMAKE_CXX_FLAGS when invoking cmake. Secondly, and less of an issue
given the platforms we are aiming for, warnings were added regardless of
the compiler in use.

This commit fixes those two issues. An ideal solution would be to
collect these in a variable and add them on a per-target basis via
target_compile_options, but that's a task for another day.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>