Add git revision information to leap-accelerate
This commit adds the necessary cmake infrastructure to calculate the git
commit SHA1 for the current build, save it into a cmake variable,
generate a C++ source file containing this information, and add it to
the leap-accelerate library. The end result is that a new git_sha1()
function is now available for querying and display, which will be useful
when inspecting tool output in the future.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>