Simplify profiling_timer class usage
This commit simplifies the profiling_timer class by removing much of its
state, and instead offering a stream output operator for easy display of
the accumulated time duration. This in turn makes it usable within log
statements or any other output streams.

Also, rather than manually stopping and resetting a timer, one simply
outputs it, and then creates a new one for the next timing operation.

Finally, all the current timing operations have now been elevated from
trace to info, as they are few and not expensive in the grand scheme of
things.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>