Avoid extra memory allocations
In PhaseMatrixFunction the A matrix and I vector, initially of a given
size, are potentially trimmed down before returning them to the user.
This trimming was implemented by allocating a new matrix and vector of
the final sizes and performing a copy from the originals, even in cases
when the original and final sizes were the same. This introduced a
close-to-instantaneous doubling of the memory usage of our process that
can be potentially problematic in the future (apart from the normal
performance implications of doing memory allocations).

Eigen's conservativeResize() method should be a no-op when no resize is
needed, and therefore we can directly apply it without further care.
Also, since we are not resizing the columns in the A matrix (only the
rows) we don't need the STATIONS variable anymore.

Finally, we move the results into the returning pair rather than copying
them in, again saving us some memory allocations.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>