Set logging level at runtime
Previously we set the logging level of our application statically
depending on the type of build we were producing. This is not very
flexible, as the logging level is intrinsically bound with the quality
of the build.

This commit adds an extra, defaulted "verbosity" argument to the logging
initialization function to indicate the desired logging level used for
filtering log records. This level is not only used for filtering logging
records, but also stored in a global variable so that users can check if
a given level is enabled *before* emitting a log message, which allows
us to have potentially expensive log statements guarded by a "if" block.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>