Move UsageReporter to its own compilation unit
Having this class sitting in a separate header and compilation unit
may remove confusion from future users who want to use this particular
class, as it wasn't immediately obvious it was contained in
resource_usage.h.

The new files are under core/profiling, and the class now sits under the
icrar::profiling namespace, as agreed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>