Include logging header correctly
The previous incarnation of this #include macro was plain wrong, but the
compilers were still able to find it -- they were probably being very
lenient.

I am still including this header *after* the system and third-party
ones as it's a "local" include. Its self-containment should be apparent
not from its inclusion into this compilation unit, but actually from its
inclusion in others where the rest of the boost headers are not
included.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>