Produce less output when downloading files
The output generated by wget is by default too verbose for
non-interactive terminals (like in Travis), which generates outputs that
are difficult to read. This commit reduces the amount of output
generated by wget by adding the "-nv" argument. I also took the
opportunity to de-duplicate some of the code.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>